gestion-documental-service/requirements.txt
pymupdf>=1.24
openai>=1.30
supabase>=2.4
thefuzz>=0.22
python-Levenshtein>=0.25
watchdog>=4.0
pydantic>=2.7
pyyaml>=6.0
Pillow>=10.3
python-dotenv>=1.0
